unroll оказывается медленнее :)

+++ sorts/insertion_sort.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include <span>
+#include <cstdint>
 
 // Time complexity: 
 // Best: O(n)
@@ -7,13 +8,13 @@
 // Worst: O(n^2)
 // Space complexity: Θ(1)
 void insertion_sort(std::span<int> arr) noexcept{
-    const size_t n = arr.size();
+    const uint64_t n = arr.size();
     if (n <= 1) return;
     int* const data = arr.data();
 
-    for (size_t i = 1; i < n; ++i){
+    for (uint64_t i = 1; i < n; ++i){
         const int x = data[i];
-        size_t j = i;
+        uint64_t j = i;
         while (j > 0 && data[j - 1] > x) data[j--] = data[j - 1];
         data[j] = x;
     }
